requirements/base.txt
# Base
pytz==2020.1                         # dependencia de time zone               
python-slugify==4.0.1
Pillow==7.2.0                        # dependencia de imagenes
psycopg2==2.8.5 --no-binary psycopg2    # es el conector con la DB postgres
               
# Django
django==3.1.1      # Django

# Environment
django-environ==0.4.5    # libreria environ

# Django REST Framework
djangorestframework==3.12.1 # el framework para crear apiREST

# Passwords security
argon2-cffi==20.1.0      # dependencia para los password security

# Static files
whitenoise==5.2.0        # dependencia para los archivos estaticos

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1
tornado>=5.0.0,<6.0.0
flower==0.9.2